% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{MinkowskiSum-imports}
\alias{MinkowskiSum-imports}
\alias{toRGL}
\alias{plotEdges}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[PolygonSoup:toRGL]{toRGL}},
  \code{\link[PolygonSoup:plotEdges]{plotEdges}}.
}
