% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-septuaginta.R
\docType{data}
\name{septuaginta}
\alias{septuaginta}
\title{Septuaginta}
\format{
A list with two fields: \code{vertices} and \code{faces}.
}
\usage{
septuaginta
}
\description{
A mesh of the Septuaginta, also known as Leonardo da Vinci's 
  72-sided sphere or Campanus's sphere. This is a convex polyhedra with 
  62 vertices and 72 faces.
}
\keyword{datasets}
